Add event listeners to numbers and operators
Established currValue, valsArr, opsArr, opsObj, and updateDisplay
-currValue is a string that appends digits as numbers are clicked
-valsArr stores numeric currVal when an operator is clicked and currVal is
cleared.
-opsArr stores the operations we need to do once equals is clicked.
The way this works is, the operator buttons' ids are tied to function names in:
-opsObj
-updateDisplay is a boolean that tells the number button event listener whether
to completely overwrite the current display or just append to it.
-number button event listeners simply add the relevant digit to currValue
  -if updateDisplay is true, the display is completely overwritten with the new digit
  and then updateDisplay is set back to false. updateDisplay will only be true directly
  after clicking an operator button.
-operator event listeners push the numeric form of the currVal into the valsArr
and then clear currVal. Then, they push the relevant funciton into opsArr based on
the id of the event target and the opsObj. Finally, they reset the updateDisplay state variable
to true.
-What remains is to tie this all together with the equals button.