Add equals button functionality
Added event listener to the equals button. First, it adds the currval
to the valsArr. Then it checks to make sure there are at least 2 values
in the valsArr (implying that there is some operation to be done).
Then, it does the first operation and pushes the result to a new array
in the function scope called interm values. Then, if there is more to be
done, a for loop goes through the operators by index and performs the operations
(stored in opsObj) with arguments intermVal[i-1] and valsArr[i+1]. After much
testing, the results appear to be accurate.
The last remaining piece of functionality is to have the calculator auto clear upon
obtaining a result and then clicking another number button before clicking clear.