Allow for continuing operations and failing to click "clear"
Added a new state variable that tracks whether a result was just calculated.
If it has been, then clicking a number button will cause a clear before updating
with the typical number button functionality. This state is turned true by the equals
button and false by other operator buttons. I also added the ability to use the result
calculated as a jumping off point for more operations. To do this, the equals button
event listener clears the ops and vals arrays and then sets the currVal to the string of the
previously calculated result. An obvious short coming at this point is that there is no decimal
button. However, I like the way the UI looks and this is not a practical application anyway,
so I will omit it. I also have not implemented any kind of rounding preference (should be as
easy as Math.Round I'd imagine.)